Spring-Boot-REST-Global-Exception-Handling
We learned to use the REST service in Spring Boot. We converted java POJO files to JSON files with Jackson. We made examples of exception handling in rest. We created endpoints with path variables. We converted exception handling to global exception handling using controller advice. This is ideal for real-time projects.